Unify mirage V3 serializers
This commit unifies V3 Mirage serializers by making all of the V3
serializers inherit from a base V3 serializer. Base V3 serializer
creates a payload based on API spec fetched from API's index
(https://api.travis-ci.org/v3). This ensures that Mirage's responses
will be much closer to reality than what we had right now.

On top of that, with API's index as a base for creating a response, it
was extremely easy to implement `include` query parameter support. So
now if a request uses `include` parameter, we will properly return a
standard representation of included records.

V3 Mirage serializer also has a mechanism similar to API's mechanism for
avoiding circular dependencies. There might be a situation when a record
embeds another record, which in turn embeds the original record. For
example a repository may embed a build (for example as `current_build`
attribute) and the build will embed the repository then. This would
normally result in a circular dependency problem (a repository embeds a
build that embeds a repository that embeds a build that embeds... etc).
In such situation API doesn't embed a full record, but only record's
@href. So the response could look like:

    {
      "@type": "repository",
      "@href": "/repository/1",
      "id": 1,
      "current_build": {
        "@type": "build",
        "@href": "/build/1",
        "id": 1,
        "repository": {
          "@href": "/repository/1"
        }
      }
    }

We mimic this mechanism in V3 serializers now.

This commit also needed some changes in V2 Mirage serializers and some
of the tests, because of the change in V3 payloads.